requirements.txt
discord-webhook==1.1.0
git+https://github.com/Merubokkusu/Discord-S.C.U.M.git#egg=discum
python-dotenv==0.21.1